fixerinos

+++ src/chatbot/models.py
@@ -2,6 +2,6 @@ from django.db import models
 
 
 class GamUser(models.Model):
-    discord_id = models.IntegerField(primary_key=True)
+    discord_id = models.BigIntegerField(primary_key=True)
     gam_coins = models.PositiveIntegerField(default=0)
     social_score = models.IntegerField(default=0)
